--- tslint.json
@@ -1,44 +1,0 @@
-{
-    "extends": [
-        "tslint:recommended"
-    ],
-    "rules": {
-        "interface-name": false,
-        "max-line-length": [
-            true,
-            100
-        ],
-        "no-console": [
-            true,
-            "time",
-            "timeEnd",
-            "trace"
-        ],
-        "no-string-literal": false,
-        "object-literal-sort-keys": false,
-        "ordered-imports": false,
-        "quotemark": [
-            true,
-            "single",
-            "avoid-escape"
-        ],
-        "trailing-comma": [
-            false,
-            {
-                "multiline": "always",
-                "singleline": "never"
-            }
-        ],
-        "variable-name": [
-            true,
-            "allow-leading-underscore",
-            "ban-keywords",
-            "check-format"
-        ],
-
-
-        "no-bitwise": false,
-        "max-classes-per-file": false,
-        "no-empty": false
-    }
-}
